pipelines/pyproject.toml
[project]
name = "worldguess"
version = "0.1.0"
description = ""
authors = [{ name = "Matheus Fillipe", email = "matheusfillipeag@gmail.com" }]
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "sqlalchemy>=2.0.42,<3",
    "python-dotenv>=1.1.1,<2",
    "psycopg2-binary>=2.9.10,<3",
    "pymemcache>=4.0.0,<5",
    "rasterio>=1.3.0,<1.4",
    "geopandas>=0.14.0,<1.1",
    "shapely>=2.1.2",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "geoalchemy2>=0.18.0",
    "joblib>=1.4.0",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1,<2",
    "ruff>=0.12.7,<0.13",
    "types-requests>=2.32.4.20250913",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["*.py"]

[tool.docformatter]
wrap-summaries = 120
wrap-descriptions = 120
in-place = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
